refactor autopagination and polling to solve issues with twitter's ajax pages
The "timeline" element constantly changes, so we cannot hold a reference to it
for too long. Fixes autopagination on replies, inbox, favorites and search.

Autopagination is done by triggering the click on the "more" button. This way
the loading of the next page is one by Twitter code, not us. We listen to
jQuery ajax events and do post-processing on the content if necessary.